create action types enum
