+++ 02-ignite-timer/.eslintrc.json
@@ -3,6 +3,7 @@
   "rules": {
     "prettier/prettier": ["error", {
       "semi": true
-    }]
+    }],
+    "no-unused-vars": "off"
   }
 }
