+++ 02-ignite-timer/src/reducers/cycles.ts
@@ -1,0 +1,62 @@
+export type Cycle = {
+  id: string;
+  task: string;
+  minutesAmount: number;
+  startDate: Date;
+  interruptedDate?: Date;
+  finishedDate?: Date;
+};
+
+interface ICyclesState {
+  cycles: Cycle[];
+  activeCycleId: string | null;
+}
+
+export enum ActionTypes {
+  ADD_NEW_CYCLE = "ADD_NEW_CYCLE",
+  INTERRUPT_CURRENT_CYCLE = "INTERRUPT_CURRENT_CYCLE",
+  MARK_CURRENT_CYCLES_AS_FINISHED = "MARK_CURRENT_CYCLES_AS_FINISHED",
+}
+
+export function cyclesReducer(state: ICyclesState, action: any) {
+  switch (action.type) {
+    case ActionTypes.ADD_NEW_CYCLE:
+      return {
+        ...state,
+        cycles: [...state.cycles, action.payload.newCycle],
+        activeCycleId: action.payload.newCycle.id,
+      };
+    case ActionTypes.INTERRUPT_CURRENT_CYCLE:
+      return {
+        ...state,
+        cycles: state.cycles.map((cycle) => {
+          if (cycle.id === state.activeCycleId) {
+            return {
+              ...cycle,
+              interruptedDate: new Date(),
+            };
+          }
+
+          return cycle;
+        }),
+        activeCycleId: null,
+      };
+    case ActionTypes.MARK_CURRENT_CYCLES_AS_FINISHED:
+      return {
+        ...state,
+        cycles: state.cycles.map((cycle) => {
+          if (cycle.id === state.activeCycleId) {
+            return {
+              ...cycle,
+              finishedDate: new Date(),
+            };
+          }
+
+          return cycle;
+        }),
+        activeCycleId: null,
+      };
+    default:
+      return state;
+  }
+}
